Move Money into transit.model.basic
