+++ src/ext-test/java/org/opentripplanner/ext/fares/impl/DefaultFareServiceTest.java
@@ -15,8 +15,8 @@ import org.junit.jupiter.api.Test;
 import org.opentripplanner.model.plan.Place;
 import org.opentripplanner.model.plan.PlanTestConstants;
 import org.opentripplanner.routing.core.FareType;
-import org.opentripplanner.routing.core.Money;
 import org.opentripplanner.transit.model._data.TransitModelForTest;
+import org.opentripplanner.transit.model.basic.Money;
 
 class DefaultFareServiceTest implements PlanTestConstants {
