+++ src/ext/java/org/opentripplanner/ext/fares/impl/AddingMultipleFareService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.routing.core.FareType;
 import org.opentripplanner.routing.core.ItineraryFares;
-import org.opentripplanner.routing.core.Money;
 import org.opentripplanner.routing.fares.FareService;
+import org.opentripplanner.transit.model.basic.Money;
 
 public class AddingMultipleFareService implements FareService, Serializable {
