+++ src/main/java/org/opentripplanner/gtfs/mapping/FareProductMapper.java
@@ -9,7 +9,7 @@ import java.util.Set;
 import org.opentripplanner.ext.fares.model.FareContainer;
 import org.opentripplanner.ext.fares.model.FareProduct;
 import org.opentripplanner.ext.fares.model.RiderCategory;
-import org.opentripplanner.routing.core.Money;
+import org.opentripplanner.transit.model.basic.Money;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 
 public class FareProductMapper {
