+++ src/main/java/org/opentripplanner/routing/core/ItineraryFares.java
@@ -12,6 +12,7 @@ import java.util.Set;
 import org.opentripplanner.ext.fares.model.FareProduct;
 import org.opentripplanner.ext.fares.model.LegProducts;
 import org.opentripplanner.model.plan.Leg;
+import org.opentripplanner.transit.model.basic.Money;
 import org.opentripplanner.util.lang.ToStringBuilder;
 
 /**
