+++ src/main/java/org/opentripplanner/transit/model/basic/LocalizedMoney.java
@@ -1,7 +1,6 @@
 package org.opentripplanner.transit.model.basic;
 
 import java.util.Locale;
-import org.opentripplanner.routing.core.Money;
 
 public record LocalizedMoney(Money money) implements I18NString {
   @Override
