+++ src/test/java/org/opentripplanner/api/mapping/FareMapperTest.java
@@ -10,7 +10,7 @@ import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.model.plan.PlanTestConstants;
 import org.opentripplanner.routing.core.FareType;
 import org.opentripplanner.routing.core.ItineraryFares;
-import org.opentripplanner.routing.core.Money;
+import org.opentripplanner.transit.model.basic.Money;
 
 class FareMapperTest implements PlanTestConstants {
