+++ src/test/java/org/opentripplanner/routing/core/MoneyTest.java
@@ -11,6 +11,7 @@ import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.opentripplanner.test.support.VariableSource;
+import org.opentripplanner.transit.model.basic.Money;
 
 class MoneyTest {
